+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 4.0.4 (unreleased)
 ------------------
 
+- Serialize if a content is folderish
+  [bloodbare]
+
+- Serialize the schema with the full behavior name
+  [bloodbare]
+
 - Upgrade to aiohttp > 3 < 4.
   Notable aiohttp changes:
     - Response.write is now a coroutine
@@ -8,6 +14,8 @@
     - Websockets send_str is now a coroutine
   [vangheem]
 
+- Dublin core should not be required
+  [bloodbare] 
 
 4.0.3 (2018-07-16)
 ------------------